🐛 image shortcode
